# src/main/resources/openapi/api.yaml
openapi: 3.1.0
info:
  title: Configuration Tracker API
  version: 0.1.0

tags:
  - name: ConfigurationTypes
    description: Manage configuration type definitions
  - name: ConfigurationChanges
    description: Track configuration value changes

paths:
  /configuration-types:
    get:
      tags: [ConfigurationTypes]
      operationId: listConfigurationTypes
      summary: List all configuration types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConfigurationTypeResponse"
    post:
      tags: [ConfigurationTypes]
      operationId: createConfigurationType
      summary: Create a new configuration type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigurationTypeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationTypeResponse"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /configuration-types/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [ConfigurationTypes]
      operationId: getConfigurationType
      summary: Get configuration type by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationTypeResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /configuration-changes:
    get:
      tags: [ConfigurationChanges]
      operationId: listConfigurationChanges
      summary: List configuration changes
      parameters:
        - name: configTypeName
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            description: "Field and direction, e.g. timestamp,desc"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationChangePage"
    post:
      tags: [ConfigurationChanges]
      operationId: createConfigurationChange
      summary: Record a new configuration change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigurationChangeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationChangeResponse"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: ConfigurationType not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /configuration-changes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [ConfigurationChanges]
      operationId: getConfigurationChange
      summary: Get configuration change by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigurationChangeResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

components:
  schemas:
    DataType:
      type: string
      enum: [STRING, NUMBER, BOOLEAN]

    ConfigurationTypeResponse:
      type: object
      required: [id, name, dataType]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        dataType:
          $ref: "#/components/schemas/DataType"

    CreateConfigurationTypeRequest:
      type: object
      required: [name, dataType]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        dataType:
          $ref: "#/components/schemas/DataType"

    ConfigurationChangeResponse:
      type: object
      required: [id, configType, previousValue, newValue, timestamp]
      properties:
        id:
          type: integer
          format: int64
        configType:
          $ref: "#/components/schemas/ConfigurationTypeResponse"
        previousValue:
          type: string
        newValue:
          type: string
        timestamp:
          type: string
          format: date-time

    ConfigurationChangePage:
      type: object
      required: [content, totalElements, totalPages, number, size]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ConfigurationChangeResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
          description: Current page number (0-based)
        size:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean

    CreateConfigurationChangeRequest:
      type: object
      required: [configTypeId, newValue]
      properties:
        configTypeId:
          type: integer
          format: int64
        previousValue:
          type: string
        newValue:
          type: string
          minLength: 1

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri